[CREATE] server package for HTTP server (#33)
