--- cmd/bruter/middleware.go
@@ -1,10 +1,0 @@
-package main
-
-import (
-	"net/http"
-)
-
-// SessionLoad loads and saves the session on every request
-func SessionLoad(next http.Handler) http.Handler {
-	return session.LoadAndSave(next)
-}
